[game] Fix a perf bug during evolution
`boards_to_work` is a Vec. That's fine until we hit the part of the evolution
loop where we're trying to see if we evolved a sibling and do
`boards_to_work.contains()`. Since it's a `Vec` that's going to be O(n).

So simply make `boards_to_work` a `HashSet` and that same line becomes O(1).

That does have the side-effect of not being able to use `chunks()` so we make a
`Vec` copy of `boards_to_work` so we can keep our chunks. I think this is a
relatively small memory trade off (a few MB) for a decent perf trade off ~10x.

Before, evolving 1.88m boards would take 25-29 minutes.
After, evolving 1.88m boards takes 2 minutes.