[determinism][board][game] Add `usable_tiles` back in, try to stabilize sorts
I notced that running with the same input/flags produces non-deterministic
results. In an attempt to fix that sort on more fields to try to get a more
stable sort result before we truncate results. The goal is to reduce the number
of "ties" so there's no wiggle room for randomness.