End of the line with this branch
In the end with the help of ChatGPT I figured out what was going on.

Since a `Board`s `id` is just a hash of its `tiles` there's lots of room for
logical hash collisions.

In some runs we would get to a board through a certain lineage, giving it a
score of X.

In other runs we would arrive to the same board through a slightly different
lineage, with a different score.

It is amazing that ChatGPT managed to figure out the rules to sorting though.
Following these rules would keep things deterministic:
1. id before cumulative_score
2. cumulative_score before evolved_via
3. evolved_via **after** evolved_from
4. tiles before usable tiles

Rule 1 meant the different lineage wouldn't matter. I think that might have been
the only one that actually mattered, though it couldn't quite tell that.